pytest_workshop/calc.py
'''
    Module for calculator functions
'''


from math import prod


class Calc:
    '''
        Calc class will support these requirements
    '''

    def add(self, *list_obj):
        '''
            Adds numbers in a list object
        '''
        try:
            res = sum(list_obj)
        except TypeError:
            return TypeError
        except Exception as unknown_e:
            return unknown_e
        else:
            return res

    def sub(self, first_parm, second_parm):
        '''
            Subtract second_parm from first_parm
        '''
        return first_parm - second_parm

    def mult(self, *list_obj):
        '''
            Multiply all elements of a list
        '''
        return prod(list_obj)
